minor refactor of integration test to remove legacy code
