+++ integration-test/src/main/scala/org/apache/livy/test/framework/Cluster.scala
@@ -75,8 +75,6 @@ trait Cluster {
 }
 
 object Cluster extends Logging {
-  private val CLUSTER_TYPE = "cluster.type"
-
   private lazy val config = {
     sys.props.get("cluster.spec")
       .filter { path => path.nonEmpty && path != "default" }
@@ -92,16 +90,13 @@ object Cluster extends Logging {
         }
         p.asScala.toMap
       }
-      .getOrElse(Map(CLUSTER_TYPE -> "mini"))
+      .getOrElse(Map.empty)
   }
 
   private lazy val cluster = {
     var _cluster: Cluster = null
     try {
-      _cluster = config.get(CLUSTER_TYPE) match {
-        case Some("mini") => new MiniCluster(config)
-        case t => throw new Exception(s"Unknown or unset cluster.type $t")
-      }
+      _cluster = new MiniCluster(config)
       Runtime.getRuntime.addShutdownHook(new Thread {
         override def run(): Unit = {
           info("Shutting down cluster pool.")
@@ -125,33 +120,4 @@ object Cluster extends Logging {
   def get(): Cluster = cluster
 
   def isRunningOnTravis: Boolean = sys.env.contains("TRAVIS")
-}
-
-trait ClusterUtils {
-
-  protected def saveProperties(props: Map[String, String], dest: File): Unit = {
-    val jprops = new Properties()
-    props.foreach { case (k, v) => jprops.put(k, v) }
-
-    val tempFile = new File(dest.getAbsolutePath() + ".tmp")
-    val out = new OutputStreamWriter(new FileOutputStream(tempFile), UTF_8)
-    try {
-      jprops.store(out, "Configuration")
-    } finally {
-      out.close()
-    }
-    tempFile.renameTo(dest)
-  }
-
-  protected def loadProperties(file: File): Map[String, String] = {
-    val in = new InputStreamReader(new FileInputStream(file), UTF_8)
-    val props = new Properties()
-    try {
-      props.load(in)
-    } finally {
-      in.close()
-    }
-    props.asScala.toMap
-  }
-
 }
